Update pagecloud key url required by blackfire
This project is pointed to by the official blackfire intergration page
dealing with ansible docs, but in implementing the steps I noticed
that the code was unable to be built by ansible due to an incorrect
key url.
See Per-repository GPG keys at https://packagecloud.io/docs
for more info.